feat(motion_detector, benchmarks, jpeg_decompressor): Minor tweak to jpeg_decompressor, add in jpeg decompression step into motion detector (required changing method names), wrote benchmarks for motion detector BREAKING CHANGE:

+++ tests/jpeg_decompressor.test.hpp
@@ -11,7 +11,7 @@
 #define JPEG_ALLOWABLE_ERROR 5
 
 struct JpegFile {
-  unsigned int filesize;
+  unsigned long filesize;
   unsigned char* data;
 };
